Update LineInPixelMatrix.java
Fix drawing of lines in the second octet.

+++ src/org/CG/infrastructure/drawings/Drawing.java
@@ -7,5 +7,5 @@ import javax.media.opengl.GL;
  * @author ldavid
  */
 public interface Drawing {
-    void draw(GL gl, float[] parameters);
+    void draw(GL gl, int[] parameters);
 }
